+++ frontend/soilhive-app/src/components/FilteringSidebar/FilteringSidebarContent/FilteringSidebarContent.tsx
@@ -1,4 +1,5 @@
 import useAvailability from 'hooks/useAvailability';
+import { useTranslation } from 'react-i18next';
 import { FilteringSidebarDataScope } from '../FilteringSidebarDataScope/FilteringSidebarDataScope';
 import { FilteringSidebarLandEcosystem } from '../FilteringSidebarLandEcosystem/FilteringSidebarLandEcosystem';
 import { FilteringSidebarParameters } from '../FilteringSidebarParameters/FilteringSidebarParameters';
@@ -6,6 +7,7 @@ import { FilteringSidebarParameters } from '../FilteringSidebarParameters/Filter
 import styles from './FilteringSidebarContent.module.scss';
 
 export function FilteringSidebarContent() {
+  const { t } = useTranslation('availability');
   const { allRasterCategories } = useAvailability();
 
   const hasRasterFilters = allRasterCategories?.some(cat => cat.id !== 'soil_groups' && cat.enabled) ?? false;
@@ -13,16 +15,16 @@ export function FilteringSidebarContent() {
   return (
     <div className={styles.FilteringSidebarContent}>
       <div data-testid="sh-filtering-sidebar-section" className={styles.Section}>
-        <p className={styles.Title}>Data scope</p>
+        <p className={styles.Title}>{t('filtering_sidebar_content.data_scope')}</p>
         <FilteringSidebarDataScope />
       </div>
       <div data-testid="sh-filtering-sidebar-section" className={styles.Section}>
-        <p className={styles.Title}>Soil parameters</p>
+        <p className={styles.Title}>{t('filtering_sidebar_content.soil_parameters')}</p>
         <FilteringSidebarParameters />
       </div>
       {hasRasterFilters && (
         <div data-testid="sh-filtering-sidebar-section" className={styles.Section}>
-          <p className={styles.Title}>LAND & ECOSYSTEM</p>
+          <p className={styles.Title}>{t('filtering_sidebar_content.land_ecosystem')}</p>
           <FilteringSidebarLandEcosystem />
         </div>
       )}