+++ frontend/soilhive-app/src/components/FilteringSidebar/FilteringSidebarHeader/FilteringSidebarHeader.tsx
@@ -2,6 +2,7 @@ import FiltersIcon from 'assets/icons/filter2-icon.svg?react';
 import CrossIcon from 'assets/icons/cross-icon.svg?react';
 import { FiltersCounter } from '../FiltersCounter/FiltersCounter';
 import useAvailability from 'hooks/useAvailability';
+import { useTranslation } from 'react-i18next';
 
 import styles from './FilteringSidebarHeader.module.scss';
 
@@ -10,17 +11,18 @@ interface Props {
 }
 
 export function FilteringSidebarHeader({ onClose }: Props) {
+  const { t } = useTranslation('availability');
   const { clearAllFilters, isFiltersSelected } = useAvailability();
   return (
     <div data-testid="sh-filtering-sidebar-header" className={styles.FilteringSidebarHeader}>
       <div className={styles.Title}>
-        <FiltersIcon /> Filters <FiltersCounter />
+        <FiltersIcon /> {t('filtering_sidebar.header')} <FiltersCounter />
       </div>
 
       <div className={styles.Actions}>
         {isFiltersSelected && (
           <span className={styles.ClearAll} data-testid="sh-filtering-sidebar-header-clear-all" role="button" onClick={clearAllFilters}>
-            Clear all
+            {t('filtering_sidebar.clear_all')}
           </span>
         )}
         <CrossIcon data-testid="sh-close-icon" role="button" onClick={onClose} />